Move ctypes error checkers into separate module
